Removed warning in mrvServer.cpp.  Updated docs.
Former-commit-id: f71c1d4e6421b5bc49fe79ac3a6293a576913743

+++ mrViewer/src/core/mrvServer.cpp
@@ -1328,7 +1328,7 @@ ConnectionUI* ViewerUI::uiConnection = NULL;
 
 void server::create(mrv::ViewerUI* ui)
 {
-   unsigned port = (unsigned) ui->uiConnection->uiServerPort->value();
+   unsigned short port = (unsigned short) ui->uiConnection->uiServerPort->value();
    ServerData* data = new ServerData;
    data->port = port;
    data->ui = ui;
